세대 별 가장 좋은 Chromosome 정보 출력 기능 추가

+++ FBNDP-MLTN/Data/DataTypes.h
@@ -238,5 +238,5 @@ struct FitnessResultData
 struct GenerationBestResultData
 {
 	FitnessResultData FitnessData;
-	uint32_t ChromosomeNumber;
+	uint64_t ChromosomeNumber;
 };
